refactor: only export what is necessary
