+++ packages/react/src/index.ts
@@ -1,5 +1,18 @@
 /* eslint-disable @typescript-eslint/triple-slash-reference */
 /// <reference path="./types/wpgraphql.d.ts" />
 
-export * from './hooks';
-export * from './provider';
+export {
+  usePosts,
+  useGeneralSettings,
+  UriInfo,
+  useUriInfo,
+  usePost,
+} from './hooks';
+
+export {
+  HeadlessProvider,
+  useApollo,
+  addApolloState,
+  QueriesConfig,
+  HeadlessProviderPageProps,
+} from './provider';
